improved accuracy of exp_accurate()
When there is cancellation in xh+xl - k*(LOG2H+LOG2Lacc),
if yl is inexact, the accuracy is ulp(yl) which gives less than 107 bits.
We reduce the accuracy of LOG2Lacc so that k*LOG2Lacc is exact,
and hopefully also yl = __builtin_fma (-k, LOG2Lacc, xl) is exact.

This reduces the total number of failures on worst cases from about 300
to about 50.