fixed bug in argument reduction
To prove that *h -= SC[i][0], we assumed that h = k*2^-55 with 0 <= k < 2^44
before, but this was wrong in the case i & 0x100, since after the fast_two_sum()
call h might no longer be an integer multiple of 2^-55.

The fix is to replace the fast_two_sum() call by *h = 0x1p-11 - *h, and
*l = -*l, which keeps the property that h is an integer multiple of 2^-55.
It does not hurt when there is a huge cancellation in 0x1p-11 - *h since
we only bound the *absolute* error, and in such a case the accurate path
will be called.