honor inexact exception for pow
This costs a few cycles: on a i7-8700, the average reciprocal throughput
increases from 51.5 cycles to 62.4 cycles.

The main idea is that we save the inexact flag before the fast path, and after
the fast path (which uses floating-point numbers, thus can produce spurious
inexact exceptions), we check if the result is exact, and if so we restore the
inexact flag.

The exactness test is_exact() mimics the code in exact_pow(), but without the
knowledge of a 117-bit accurate approximation of x^y. Since it is in the
critical path, we have tests in is_exact() to help early detection of cases
that can't be exact. Maybe these tests can be improved.