added macro SKIP_C_FUNC_REDEF
Some functions are not available in GNU libc, and in that case
alternate implementations are defined.

But this creates issues when another library than GNU libc is
tested (for example the Intel library or LLVM).

When this macro is defined (whose name was suggested by Tue Ly),
the redefinition of the alternate functions is skipped.