[pow] try to fix ERANGE issues
This is not yet in final state. Indeed, the C standard says that ERANGE
should be set when overflow occurs. IEEE 754-2019 says that overflow
occurs when the rounded result with unbounded exponent is not representable
in the given format as a regular number (thus overflow depends on the
rounding mode).

This means that we cannot have simply lines like "x,y,ERANGE" in pow.wc,
we should also give the rounding modes for which overflow occurs.