perf_common.sh: Add __rdtsc powerpc support
The 'mfspr 256' access the time base value (Book 2, Chapter 5 from
PowerISA) and it is avaliable since ISA 2.01 (power4). The suggested
frequency at which the time base increments is 512 MHz, however the
ABI allow variations.  The glibc provides the function
__ppc_get_timebase_freq, which allows to query the timebase frequency
using either the vDSO or by parsing the procfs.

Similar to aarch64, to get accurate measurements a lwsync barrier is
added prior mfspr.